Оставил комментарии. Оптимизировал код. Немного поменял структуру в первом задании.

+++ task_1/login_app/server_handler/main.py
@@ -1,22 +1,24 @@
 import json
+
+from flask import Blueprint, render_template, redirect, url_for
 from flask import request
-from flask import Blueprint, current_app, render_template, redirect, url_for
 from flask_login import login_required, current_user
 from werkzeug.security import generate_password_hash
 
 from login_app import db
-from login_app.first.forms import ChangeUserForm
-from login_app.first.models import Role, MyUsers
+from login_app.forms import ChangeUserForm
+from login_app.models import Role, MyUsers
 
 main = Blueprint('main', __name__)
 
 
-def log_error(*args, **kwargs):
-    current_app.logger.error(*args, **kwargs)
-
-
 @main.route('/', methods=['GET'])
 def index():
+    """
+    Рендер главной страницы. Доступна только "залогинившимся" пользователям.
+    :return:
+    """
+    # Если пользователь не вошел в систему, редирект на страницу "логина"
     if not current_user.is_authenticated:
         return redirect(url_for('auth.login'))
     all_user = MyUsers.query.all()
@@ -26,23 +28,37 @@ def index():
 @main.route('/delete_user', methods=["POST"])
 @login_required
 def change_user_post_delete():
-    data = json.loads(request.data.decode('utf-8'))
-    user = MyUsers.query.get(int(data['id']))
-    db.session.delete(user)
-    db.session.commit()
-    return json.dumps({'msg': 'Пользователь успешно удален!'})
+    """
+    Удаление пользователя из базы данных. Вход через AJAX.
+    :return:
+    """
+    # "Защита" от отправки сигналов из postman и прочих вещей, чтобы пользователей не мог удалить кто угодно.
+    if current_user.roles[0].name == 'Админ' or current_user.roles[0].name == 'Редактор':
+        data = request.json
+        user = MyUsers.query.get(int(data['id']))
+        db.session.delete(user)
+        db.session.commit()
+        return json.dumps({'msg': 'Пользователь успешно удален!'})
+    else:
+        return json.dumps({'msg': 'А кто это тут пытается у меня пользователей удалить, а?)))'})
 
 
 @main.route('/change_user', methods=["POST"])
 @login_required
 def change_user_post_change():
-    data = json.loads(request.data.decode('utf-8'))
+    """
+    Редактирование пользователей. Вход из AJAX
+    :return:
+    """
+    data = request.json
     user = MyUsers.query.get(int(data['id']))
     user.username = data['username']
     user.email = data['email']
     user.roles[0] = (Role.query.get(int(data['role'])))
+    # если пароль введен, проверка валидности пароля для замены
     if data['cur_pass']:
         if data['pass1'] == data['pass2']:
+            # Если текущий пароль введен верно, смена пароля на новый
             if user.check_password(data['cur_pass']):
                 user.password = generate_password_hash(data['pass1'])
             else:
@@ -57,6 +73,11 @@ def change_user_post_change():
 @main.route('/change_user/<int:pk>', methods=["GET"])
 @login_required
 def change_user_get_change(pk):
+    """
+    Рендер страницы редактирования пользователя
+    :param pk:
+    :return:
+    """
     form = ChangeUserForm()
     user = MyUsers.query.get(int(pk))
     form.username.render_kw['value'] = user.username
@@ -70,6 +91,8 @@ def change_user_get_change(pk):
             if role.id != user.roles[0].id:
                 form.role.choices.append((role.id, role))
     except IndexError:
+        # Обезопасил момент, когда у пользователя нет никакой роли(по умолчанию новые пользователи получают
+        # роль "пользователь")
         form.role.choices = [('-', '-')]
         for role in all_roles:
             form.role.choices.append((role.id, role))
@@ -77,8 +100,12 @@ def change_user_get_change(pk):
 
 
 @main.route('/change_role', methods=['GET'])
-def change_roles():
+def get_roles():
+    """
+    Рендер страницы со списком всех ролей. Доступно только суперпользователю.
+    :return:
+    """
     all_roles = Role.query.all()
     if not current_user.is_superuser:
         return redirect(url_for('main.index'))
-    return render_template('change_roles.html', roles=all_roles)
+    return render_template('get_roles.html', roles=all_roles)
